+++ ddb/feature/gitignore/schema.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from marshmallow import fields
 
 from ddb.feature.schema import FeatureSchema
 
@@ -7,3 +8,4 @@ class GitignoreSchema(FeatureSchema):
     """
     Git schema.
     """
+    enforce = fields.List(fields.String(), default=["*ddb.local.*"])